simulator/requirements.txt
fastapi
uvicorn
pydantic
numpy
torch
Pillow
python-dotenv
azure-identity
azure-digitaltwins-core
